Move upload logic to add-annotation.component
