+++ src/app/components/import-annotations/details/details.component.ts
@@ -7,24 +7,14 @@ import {
   AudioEventImportService,
   audioEventImportResolvers,
 } from "@baw-api/audio-event-import/audio-event-import.service";
-import { contains, filterAnd, notIn } from "@helpers/filters/filters";
-import { Tag } from "@models/Tag";
 import { takeUntil, Observable, BehaviorSubject } from "rxjs";
 import { Id } from "@interfaces/apiInterfaces";
 import { AudioEvent } from "@models/AudioEvent";
-import { TagsService } from "@baw-api/tag/tags.service";
-import { AudioEventImportFileWrite } from "@models/AudioEventImport/AudioEventImportFileWrite";
 import { Filters } from "@baw-api/baw-api.service";
-import { AbstractModel } from "@models/AbstractModel";
 import { defaultSuccessMsg } from "@helpers/formTemplate/formTemplate";
 import { ToastrService } from "ngx-toastr";
 import { ShallowAudioEventsService } from "@baw-api/audio-event/audio-events.service";
-import { BawApiError } from "@helpers/custom-errors/baw-api-error";
-import { FORBIDDEN } from "http-status";
-import {
-  AudioEventError,
-  ImportedAudioEvent,
-} from "@models/AudioEventImport/ImportedAudioEvent";
+import { ImportedAudioEvent } from "@models/AudioEventImport/ImportedAudioEvent";
 import { AudioEventImportFile } from "@models/AudioEventImportFile";
 import { AudioEventImportFileService } from "@baw-api/audio-event-import-file/audio-event-import-file.service";
 import { deleteAnnotationImportModal } from "../import-annotations.modals";
@@ -68,7 +58,6 @@ interface ImportGroup {
 class AnnotationsDetailsComponent extends PageComponent implements OnInit {
   public constructor(
     private route: ActivatedRoute,
-    private tagsApi: TagsService,
     private eventsApi: ShallowAudioEventsService,
     private eventImportsApi: AudioEventImportService,
     private eventImportFileApi: AudioEventImportFileService,
@@ -161,202 +150,6 @@ class AnnotationsDetailsComponent extends PageComponent implements OnInit {
         },
       });
   }
-
-  // since the typeahead input returns an array of models, but the api wants the associated tags as an array of id's
-  // we use this helper function to convert the array of models to an array of id's that can be sent in the api request
-  protected getIdsFromAbstractModelArray(items: object[]): Id[] {
-    return items.map((item: AbstractModel): Id => item.id);
-  }
-
-  protected pushToImportGroups(model: ImportGroup, event): void {
-    const files: FileList = event.target.files;
-    model.files = files;
-
-    this.performDryRun(model);
-
-    // if the user updates an existing import group, we don't want to create a new one
-    // however, if the user uses the last empty import group, we want to create a new empty one
-    // that they can use to create a new import group
-    if (this.areImportGroupsFull()) {
-      this.importGroups.push(this.emptyImportGroup);
-    }
-  }
-
-  // uses a reference to the ImportGroup object and update the additional tag ids property
-  protected updateAdditionalTagIds(
-    model: ImportGroup,
-    additionalTagIds: Id[]
-  ): void {
-    model.additionalTagIds = additionalTagIds;
-    this.performDryRun(model);
-  }
-
-  // a predicate to check if every import group is valid
-  // this is used for form validation
-  protected areImportGroupsValid(): boolean {
-    const importErrors = this.importGroups.flatMap((model) => model?.errors);
-    return this.importGroups.length > 1 && importErrors.length === 0;
-  }
-
-  protected removeFromImport(model: ImportGroup): void {
-    const index = this.importGroups.indexOf(model);
-    if (index !== -1) {
-      this.importGroups.splice(index, 1);
-    }
-  }
-
-  // sends all import groups to the api if there are no errors
-  protected async uploadImportGroups(): Promise<void> {
-    // importing invalid annotation imports results in an internal server error
-    // we should therefore not submit any upload groups if there are any errors
-    if (!this.areImportGroupsValid()) {
-      return;
-    }
-
-    // creates a lock so that no more files can be added to the upload queue while the upload is in progress
-    this.uploading = true;
-
-    // we use a "for-of" loop here because if we use a forEach loop (with async callbacks)
-    // it doesn't properly await for each import group to finish uploading
-    for (const model of this.importGroups) {
-      await this.importEventGroup(model);
-    }
-
-    this.importGroups = this.importGroups.filter((model) => !model.uploaded);
-
-    this.refreshTables();
-    this.uploading = false;
-  }
-
-  protected async importEventGroup(model: ImportGroup): Promise<void> {
-    if (!model.files) {
-      return;
-    }
-
-    for (const file of model.files) {
-      await this.uploadFile(model, file);
-    }
-  }
-
-  private uploadFile(
-    model: ImportGroup,
-    file: File
-  ): Promise<void | AudioEventImport> {
-    const audioEventImportModel: AudioEventImportFileWrite =
-      new AudioEventImportFileWrite({
-        id: this.audioEventImport.id,
-        file,
-        additionalTagIds: model.additionalTagIds,
-        commit: true,
-      });
-
-    return this.eventImportsApi
-      .importFile(audioEventImportModel)
-      .pipe(takeUntil(this.unsubscribe))
-      .toPromise()
-      .finally(() => {
-        model.uploaded = true;
-      })
-      .catch((error: BawApiError) => {
-        // some of the default error messages are ambiguous on this page
-        // e.g. "you do not have access to this page" means that the user doesn't have access to the audio recording
-        if (error.status === FORBIDDEN) {
-          model.errors.push(
-            "You do not have access to all the audio recordings or tags in your files."
-          );
-        }
-      });
-  }
-
-  private performDryRun(model: ImportGroup) {
-    model.errors = [];
-    model.identifiedEvents = [];
-
-    // we perform a dry run of the import to check for errors
-    for (const file of model.files) {
-      const audioEventImportModel: AudioEventImportFileWrite =
-        new AudioEventImportFileWrite({
-          id: this.audioEventImport.id,
-          file,
-          additionalTagIds: model.additionalTagIds,
-          commit: false,
-        });
-
-      this.eventImportsApi
-        .importFile(audioEventImportModel)
-        .pipe(takeUntil(this.unsubscribe))
-        .subscribe((result: AudioEventImport) => {
-          // since the model is on the heap and passed as reference, we can update the model here and it will globally update the model
-          result.importedEvents.forEach((event: ImportedAudioEvent) => {
-            model.identifiedEvents.push(event);
-
-            const errors: AudioEventError[] = event.errors;
-
-            for (const error of errors) {
-              model.errors.push(...this.errorToHumanReadable(error));
-            }
-          });
-        });
-    }
-  }
-
-  // deserialization converts all object keys to camelCase
-  // therefore, to make them human readable we add spaces
-  private errorToHumanReadable(error: AudioEventError): string[] {
-    const errors: string[] = [];
-
-    const errorKeys: string[] = Object.keys(error);
-
-    for (const errorKey of errorKeys) {
-      let errorValue: string = error[errorKey].join(", ");
-
-      // sometimes the error value includes the key e.g. "startTimeSeconds is not a number"
-      // in this case, we should not prepend the key to the error value
-      if (!errorValue.includes(errorKey)) {
-        errorValue = `${errorKey} ${errorValue}`;
-      }
-
-      errors.push(errorValue);
-    }
-
-    return errors;
-  }
-
-  // because we create a new empty import group if all import groups are full
-  // we use this predicate to check if every import group has files
-  private areImportGroupsFull(): boolean {
-    return this.importGroups.every((model) => model.files !== null);
-  }
-
-  // because the event and file tables are updated through api requests
-  // we have to make new api requests to the page
-  private refreshTables(): void {
-    // we use the default filter here to prevent two api requests being sent
-    // e.g. if we set the filters to {} here, it would make an api call with the {} filters
-    // then another one with the default filters
-    this.eventFilters$.next(this.defaultEventFilters);
-
-    // because the "files" property is a sub model on the audio event import model
-    // we have to refetch the audio event import model to update the files table
-    this.eventImportsApi
-      .show(this.audioEventImport)
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: AudioEventImport) => {
-        this.audioEventImport = result;
-      });
-  }
-
-  // callback used by the typeahead input to search for associated tags
-  protected searchTagsTypeaheadCallback = (
-    text: string,
-    activeItems: Tag[]
-  ): Observable<Tag[]> =>
-    this.tagsApi.filter({
-      filter: filterAnd(
-        contains<Tag, "text">("text", text),
-        notIn<Tag>("text", activeItems)
-      ),
-    });
 }
 
 AnnotationsDetailsComponent.linkToRoute({
